Refactor skills page components
